// app/src/types/modules/financial-maintenance/discount-management/DiscountManagementTypes.ts
export type DiscountType = "Percentage" | "Fixed";
export type DiscountTransactionType = "Purchase" | "Sales";
export type DiscountStatus = "Active" | "Inactive";
// ...
export type DiscountManagementFormValues = {
	name: string;
	description: string;
	type: DiscountTransactionType;
	discountType: DiscountType;
	amount: string;
	status: DiscountStatus;
};

export type DiscountManagementFormErrors = Partial<
	Record<keyof DiscountManagementFormValues, string>
>;